vlmeval/vlm/valley/requirements_valley.txt
fastapi==0.115.0
requests==2.32.3
protobuf==3.20.3
uvicorn==0.30.6
markdown2==2.5.0
numpy==1.26.4
ninja==1.11.1.1
decord==0.6.0
scikit-image==0.24.0
opencv-python-headless==4.10.0.84
prettytable==3.11.0
accelerate==0.34.2
einops==0.8.0
sentencepiece==0.1.97
tokenizers>=0.13.3
evaluate==0.4.3
bert-score==0.3.13
rouge-score==0.1.2
datasets==2.21.0
packaging==24.1
nltk==3.9.1
scikit-learn==1.5.2
pandas==2.2.2
byted-wandb==0.13.72
peft==0.5.0
deepspeed==0.9.5
openai==0.28
omegaconf==2.3.0
pyarrow==15.0.0
timm==0.6.7
ftfy==6.2.3
pydantic==1.10.14
transformers==4.45.2
torchmetrics
qwen_vl_utils
flash_attn
